+++ common/src/main/java/nl/nielspoldervaart/gdmc/common/utils/CustomHeightmap.java
@@ -1,5 +1,6 @@
 package nl.nielspoldervaart.gdmc.common.utils;
 
+import net.minecraft.commands.arguments.blocks.BlockStateParser;
 import net.minecraft.core.BlockPos;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.tags.TagKey;
@@ -44,8 +45,8 @@ public class CustomHeightmap {
 		this.yMaxBound = yMaxBound;
 	}
 
-	public static CustomHeightmap primeHeightmaps(ChunkAccess chunk, ArrayList<BlockState> blockList, ArrayList<String> blockTagLocationKeyList, Optional<Integer> yMinBound, Optional<Integer> yMaxBound) {
-		Predicate<BlockState> isOpaque = blockState -> !blockList.contains(blockState) && !hasBlockTagKey(blockState, blockTagLocationKeyList);
+	public static CustomHeightmap primeHeightmaps(ChunkAccess chunk, ArrayList<BlockStateParser.BlockResult> blockStateParserList, ArrayList<String> blockTagLocationKeyList, Optional<Integer> yMinBound, Optional<Integer> yMaxBound) {
+		Predicate<BlockState> isOpaque = blockState -> !hasBlockState(blockState, blockStateParserList) && !hasBlockTagKey(blockState, blockTagLocationKeyList);
 		CustomHeightmap customHeightmap = new CustomHeightmap(chunk, isOpaque, yMinBound, yMaxBound);
 		return primeHeightmaps(chunk, customHeightmap);
 	}
@@ -56,8 +57,21 @@ public class CustomHeightmap {
 	}
 
 	private static CustomHeightmap primeHeightmaps(ChunkAccess chunk, CustomHeightmap customHeightmap) {
-		int yMax = customHeightmap.yMaxBound.orElse(getMaxY(chunk));
-		int yMin = customHeightmap.yMinBound.orElse(getMinY(chunk));
+		// Cannot use Math.clamp here, since that isn't yet supported in Java 17.
+		int yMin = Math.max(
+			Math.min(
+				customHeightmap.yMinBound.orElse(getChunkMinY(chunk)),
+				getChunkMaxY(chunk)
+			),
+			getChunkMinY(chunk)
+		);
+		int yMax = Math.max(
+			Math.min(
+				customHeightmap.yMaxBound.orElse(getChunkMaxY(chunk)),
+				getChunkMaxY(chunk)
+			),
+			getChunkMinY(chunk)
+		);
 		BlockPos.MutableBlockPos blockPos = new BlockPos.MutableBlockPos();
 		for (int x = 0; x < 16; x++) {
 			for (int z = 0; z < 16; z++) {
@@ -75,7 +89,7 @@ public class CustomHeightmap {
 	}
 
 	private void setHeight(int x, int z, int y) {
-		this.data.set(getIndex(x, z), y - getMinY(this.chunk));
+		this.data.set(getIndex(x, z), y - getChunkMinY(this.chunk));
 	}
 
 	public int getFirstAvailable(int x, int z) {
@@ -83,14 +97,20 @@ public class CustomHeightmap {
 	}
 
 	private int getFirstAvailable(int index) {
-		return this.data.get(index) + getMinY(this.chunk);
+		return this.data.get(index) + getChunkMinY(this.chunk);
 	}
 
 	private static int getIndex(int x, int z) {
 		return x + z * 16;
 	}
 
-	private static int getMinY(ChunkAccess chunk) {
+	/**
+	 * Get min Y value of vertical world limit at given chunk
+	 *
+	 * @param chunk     chunk to check minimum height of
+	 * @return          min Y value
+	 */
+	private static int getChunkMinY(ChunkAccess chunk) {
 		#if (MC_VER == MC_1_21_4)
 		return chunk.getMinY();
 		#else
@@ -98,12 +118,30 @@ public class CustomHeightmap {
 		#endif
 	}
 
-	private static int getMaxY(ChunkAccess chunk) {
+	/**
+	 * Get max Y value of vertical world limit at given chunk
+	 *
+	 * @param chunk     chunk to check maximum height of
+	 * @return          max Y value
+	 */
+	private static int getChunkMaxY(ChunkAccess chunk) {
 		#if (MC_VER == MC_1_21_4)
 		return chunk.getMaxY();
 		#else
 		return chunk.getMaxBuildHeight();
 		#endif
+	}
+
+	private static boolean hasBlockState(BlockState blockState, ArrayList<BlockStateParser.BlockResult> inputBlockStateParserList) {
+		if (inputBlockStateParserList.isEmpty()) {
+			return false;
+		}
+		return inputBlockStateParserList.stream().anyMatch(inputBlockStateParserResult -> {
+			if (inputBlockStateParserResult.properties().isEmpty()) {
+				return blockState.getBlock().defaultBlockState().equals(inputBlockStateParserResult.blockState());
+			}
+			return blockState.equals(inputBlockStateParserResult.blockState());
+		});
 	}
 
 	private static boolean hasBlockTagKey(BlockState blockState, ArrayList<String> inputBlockTagLocationKeyList) {